requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
scikit-learn
keras
pandas
python-utils
typing